+++ kie-dmn/kie-dmn-core-jsr223/src/main/java/org/kie/dmn/core/jsr223/JSR223LiteralExpressionEvaluator.java
@@ -1,0 +1,33 @@
+package org.kie.dmn.core.jsr223;
+
+import org.kie.dmn.api.core.DMNResult;
+import org.kie.dmn.api.core.event.DMNRuntimeEventManager;
+import org.kie.dmn.core.api.DMNExpressionEvaluator;
+import org.kie.dmn.core.api.EvaluatorResult;
+import org.kie.dmn.core.api.EvaluatorResult.ResultType;
+import org.kie.dmn.core.ast.EvaluatorResultImpl;
+import org.kie.dmn.core.impl.DMNResultImpl;
+
+public class JSR223LiteralExpressionEvaluator implements DMNExpressionEvaluator {
+    
+    private final JSR223ScriptEngineEvaluator eval;
+
+    public JSR223LiteralExpressionEvaluator(JSR223ScriptEngineEvaluator eval) {
+        this.eval = eval;
+    }
+
+    @Override
+    public EvaluatorResult evaluate(DMNRuntimeEventManager eventManager, DMNResult dmnr) {
+        DMNResultImpl result = (DMNResultImpl) dmnr;
+        Object evaluatorResult = null;
+        ResultType resultType = ResultType.SUCCESS;
+        try {
+            evaluatorResult = eval.eval(result.getContext().getAll());
+        } catch (Exception e) {
+            e.printStackTrace();
+            resultType = ResultType.FAILURE;
+        }
+        return new EvaluatorResultImpl(evaluatorResult, resultType);
+    }
+
+}
